+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix Logic
 - Fix Bug
 - Fixed Ping CURL / HTTP / HTTPS
+- Fixed Custom Script
 
 **Changelog** (v0.00.23-beta 19-05-2024):
 - Perbaikan Ping Di HTTPS Dan HTTP